+++ lib/hasura_connect.dart
@@ -7,3 +7,4 @@ export 'src/domain/entities/snapshot.dart';
 export 'src/domain/errors/errors.dart';
 export 'src/domain/models/request.dart';
 export 'src/presenter/hasura_connect_base.dart';
+export 'src/domain/models/query.dart';